requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
